replaced unsigned int with size_t

+++ Arrangement_on_surface_2/include/CGAL/Sweep_line_2/Arr_construction_sl_visitor.h
@@ -314,7 +314,7 @@ after_handle_event(Event* event, Status_line_iterator iter, bool /* flag */)
   if (event->parameter_space_in_x() == CGAL::ARR_LEFT_BOUNDARY) {
     if (!this->is_status_line_empty()) {
       Status_line_iterator prev = iter;
-      for (unsigned i = 0; i < event->number_of_right_curves(); ++i) --prev;
+      for (size_t i = 0; i < event->number_of_right_curves(); ++i) --prev;
       // move items from top face to last inserted curve
       Indices_list& list_ref = (*prev)->halfedge_indices_list();
       list_ref.clear();
